xtask: inline build_ebpf
